pyproject.toml
[tool.poetry]
name = "skatepark"
version = "0.7.0"
description = "Python framework for AI workflows and pipelines."
authors = ["Griptape <hello@griptape.ai>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/skatepark"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = ">=0.21"
openai = ">=0.27"
attrs = ">=22"
jinja2 = ">=3.1"
jsonschema = ">=4"
marshmallow = ">=3"
marshmallow-enum = ">=1.5"
graphlib = "*"
llama_index = "0.4.23"
wikipedia = "*"
tiktoken = ">=0.3"
gspread = "*"
sqlalchemy = ">1"
rich = ">=13"
trafilatura = "*"
requests = "*"
warpspeed-googlesearch-python = ">=1.1.1"
stopit = "*"

[tool.poetry.group.test.dependencies]
pytest = "~=7.1"
pytest-cover = "*"
twine = ">=4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
